pyproject.toml
[tool.poetry]
name = "psion"
version = "0.1.0"
description = "Asynchronous WebAuth library for python."
authors = ["Eduardo Ribeiro Rezende <eduardorbr7@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/revensky/psion"
repository = "https://github.com/revensky/psion.git"
keywords = ["JOSE", "OAuth 2.1", "OpenID Connect"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
packages = [{ include = "psion", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
cryptography = "^3.3.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^20.8b1"
Sphinx = "^3.4.3"
rstcheck = "^3.3.1"
sphinx-rtd-theme = "^0.5.1"
starlette = "^0.14.2"
email-validator = "^1.1.2"
requests = "^2.25.1"
aiofiles = "^0.6.0"
Jinja2 = "^2.11.3"
python-multipart = "^0.0.5"
itsdangerous = "^1.1.0"
Starlette-WTF = "^0.3.0"
tortoise-orm = "^0.16.21"
marshmallow = "^3.10.0"
python-dotenv = "^0.15.0"
asyncpg = "^0.21.0"
asynctest = "^0.13.0"
uvicorn = "^0.13.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
